+++ src/configuration/read-configuration.ts
@@ -27,6 +27,8 @@ import {
   getInactiveBackgroundColorHex,
   getInactiveForegroundColorHex,
 } from '../color-library';
+import { RemoteSettings } from '../remote/enums';
+import { LiveShareSettings } from '../live-share/enums';
 
 const { workspace } = vscode;
 
@@ -105,6 +107,14 @@ export function prepareColors(backgroundHex: string) {
 
 export function getDarkForegroundColor() {
   return readConfiguration<string>(StandardSettings.DarkForegroundColor, '');
+}
+
+export function getRemoteColor(remoteSetting: RemoteSettings) {
+  return readConfiguration<string>(remoteSetting, '');
+}
+
+export function getLiveShareColor(liveShareSetting: LiveShareSettings) {
+  return readConfiguration<string>(liveShareSetting, '');
 }
 
 export function getDarkForegroundColorOrOverride() {