+++ src/live-share/constants.ts
@@ -1,7 +1,0 @@
-/* istanbul ignore file */
-import { extensionShortName } from '../models';
-
-export const peacockVslsMementos = {
-  vslsShareColor: `${extensionShortName}.vsls.share.color`,
-  vslsJoinColor: `${extensionShortName}.vsls.join.color`,
-};