+++ src/live-share/enums.ts
@@ -3,3 +3,8 @@ export enum LiveShareCommands {
   changeColorOfLiveShareHost = 'peacock.changeColorOfLiveShareHost',
   changeColorOfLiveShareGuest = 'peacock.changeColorOfLiveShareGuest',
 }
+
+export enum LiveShareSettings {
+  VSLSShareColor = 'vslsShareColor',
+  VSLSJoinColor = 'vslsJoinColor',
+}