+++ src/live-share/test/suite/live-share.test.ts
@@ -3,7 +3,7 @@ import sinon = require('sinon');
 import assert = require('assert');
 import * as vsls from 'vsls';
 
-import { IPeacockSettings, Commands, ColorSettings, timeout } from '../../../models';
+import { IPeacockSettings, Commands, ColorSettings, timeout, azureBlue } from '../../../models';
 import {
   setupTestSuite,
   teardownTestSuite,
@@ -11,9 +11,12 @@ import {
 } from '../../../test/suite/lib/setup-teardown-test-suite';
 import { isValidColorInput } from '../../../color-library';
 import { executeCommand } from '../../../test/suite/lib/constants';
-import { LiveShareCommands } from '../../enums';
-import { peacockVslsMementos } from '../../constants';
-import { getPeacockWorkspaceConfig } from '../../../configuration';
+import { LiveShareCommands, LiveShareSettings } from '../../enums';
+import {
+  getPeacockWorkspaceConfig,
+  updateLiveShareColor,
+  getLiveShareColor,
+} from '../../../configuration';
 
 suite('Live Share Integration', () => {
   let originalValues = <IPeacockSettings>{};
@@ -23,47 +26,27 @@ suite('Live Share Integration', () => {
   setup(async () => await setupTest());
 
   test('can set color setting for Live Share host', async () => {
-    // Stub the async quick pick to return a response
-    const color = '#007fff';
-    const fakeResponse = `Azure Blue -> ${color}`;
-    const stub = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse));
-
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
-      LiveShareCommands.changeColorOfLiveShareHost,
-    );
-
-    const settingValue =
-      (await extensionContext!.globalState.get<string>(peacockVslsMementos.vslsShareColor)) || '';
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsShareColor, null);
-
+    const fakeResponse = `Azure Blue -> ${azureBlue}`;
+    const stub = await stubQuickPick(fakeResponse);
+    await executeCommand(LiveShareCommands.changeColorOfLiveShareHost);
+    const settingValue = getLiveShareColor(LiveShareSettings.VSLSShareColor) || '';
+    await updateLiveShareColor(LiveShareSettings.VSLSShareColor, '');
     stub.restore();
 
-    assert(isValidColorInput(settingValue));
-    assert(settingValue === color);
+    assert.ok(isValidColorInput(settingValue));
+    assert.equal(settingValue, azureBlue);
   });
 
   test('can set color setting for Live Share guest', async () => {
-    // Stub the async quick pick to return a response
-    const color = '#007fff';
-    const fakeResponse = `Azure Blue -> ${color}`;
-    const stub = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse));
-
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
-      LiveShareCommands.changeColorOfLiveShareGuest,
-    );
-
-    const settingValue =
-      (await extensionContext!.globalState.get<string>(peacockVslsMementos.vslsJoinColor)) || '';
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsJoinColor, null);
-
+    const fakeResponse = `Azure Blue -> ${azureBlue}`;
+    const stub = await stubQuickPick(fakeResponse);
+    await executeCommand(LiveShareCommands.changeColorOfLiveShareGuest);
+    const settingValue = getLiveShareColor(LiveShareSettings.VSLSJoinColor) || '';
+    await updateLiveShareColor(LiveShareSettings.VSLSJoinColor, '');
     stub.restore();
 
-    assert(isValidColorInput(settingValue));
-    assert(settingValue === color);
+    assert.ok(isValidColorInput(settingValue));
+    assert.equal(settingValue, azureBlue);
   });
 
   test('Workspace color is updated when Live Share session is started.', async () => {
@@ -73,15 +56,9 @@ suite('Live Share Integration', () => {
       throw new Error('Live Share extension is not installed.');
     }
 
-    const color = '#007fff';
-    const fakeResponse = `Azure Blue -> ${color}`;
-    const stub = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse));
-
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
-      LiveShareCommands.changeColorOfLiveShareHost,
-    );
+    const fakeResponse = `Azure Blue -> ${azureBlue}`;
+    const stub = await stubQuickPick(fakeResponse);
+    await executeCommand(LiveShareCommands.changeColorOfLiveShareHost);
     stub.restore();
 
     await vslsApi.share();
@@ -91,10 +68,10 @@ suite('Live Share Integration', () => {
     const value = config[ColorSettings.titleBar_activeBackground];
 
     await vslsApi.end();
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsShareColor, null);
+    await updateLiveShareColor(LiveShareSettings.VSLSShareColor, '');
 
-    assert(isValidColorInput(value));
-    assert(value === color);
+    assert.ok(isValidColorInput(value));
+    assert.equal(value, azureBlue);
   });
 
   test('Workspace color is reverted when Live Share session is ended.', async () => {
@@ -104,15 +81,9 @@ suite('Live Share Integration', () => {
       throw new Error('Live Share extension is not installed.');
     }
 
-    const color = '#007fff';
-    const fakeResponse = `Azure Blue -> ${color}`;
-    const stub = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse));
-
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
-      LiveShareCommands.changeColorOfLiveShareHost,
-    );
+    const fakeResponse = `Azure Blue -> ${azureBlue}`;
+    const stub = await stubQuickPick(fakeResponse);
+    await executeCommand(LiveShareCommands.changeColorOfLiveShareHost);
     stub.restore();
 
     await vslsApi.share();
@@ -123,10 +94,10 @@ suite('Live Share Integration', () => {
     let config = getPeacockWorkspaceConfig();
     const value = config[ColorSettings.titleBar_activeBackground];
 
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsShareColor, null);
+    await updateLiveShareColor(LiveShareSettings.VSLSShareColor, '');
 
-    assert(!isValidColorInput(value));
-    assert(value == null);
+    assert.ok(!isValidColorInput(value));
+    assert.ok(!value);
   });
 
   test('Workspace color is immediately reflected when set during Live Share session.', async () => {
@@ -138,25 +109,19 @@ suite('Live Share Integration', () => {
 
     await vslsApi.share();
 
-    const color = '#007fff';
-    const fakeResponse = `Azure Blue -> ${color}`;
-    const stub = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse));
-
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
-      LiveShareCommands.changeColorOfLiveShareHost,
-    );
+    const fakeResponse = `Azure Blue -> ${azureBlue}`;
+    const stub = await stubQuickPick(fakeResponse);
+    await executeCommand(LiveShareCommands.changeColorOfLiveShareHost);
     stub.restore();
 
     let config = getPeacockWorkspaceConfig();
     const value = config[ColorSettings.titleBar_activeBackground];
 
     await vslsApi.end();
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsShareColor, null);
+    await updateLiveShareColor(LiveShareSettings.VSLSShareColor, '');
 
-    assert(isValidColorInput(value));
-    assert(value === color);
+    assert.ok(isValidColorInput(value));
+    assert.equal(value, azureBlue);
   });
 
   test('Workspace color is immediately reflected when updated during Live Share session.', async () => {
@@ -166,15 +131,9 @@ suite('Live Share Integration', () => {
       throw new Error('Live Share extension is not installed.');
     }
 
-    const color = '#007fff';
-    const fakeResponse = `Azure Blue -> ${color}`;
-    const stub = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse));
-
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
-      LiveShareCommands.changeColorOfLiveShareHost,
-    );
+    const fakeResponse = `Azure Blue -> ${azureBlue}`;
+    const stub = await stubQuickPick(fakeResponse);
+    await executeCommand(LiveShareCommands.changeColorOfLiveShareHost);
     stub.restore();
 
     await vslsApi.share();
@@ -183,11 +142,8 @@ suite('Live Share Integration', () => {
 
     const color2 = '#68217a';
     const fakeResponse2 = `C# Purple -> ${color2}`;
-    const stub2 = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse2));
-
-    await executeCommand<vscode.ExtensionContext>(LiveShareCommands.changeColorOfLiveShareHost);
+    const stub2 = await stubQuickPick(fakeResponse2);
+    await executeCommand(LiveShareCommands.changeColorOfLiveShareHost);
     stub2.restore();
 
     await timeout(1000);
@@ -195,11 +151,11 @@ suite('Live Share Integration', () => {
     let config = getPeacockWorkspaceConfig();
     const value = config[ColorSettings.titleBar_activeBackground] as string;
 
-    assert(isValidColorInput(value));
-    assert(value === color2);
+    assert.ok(isValidColorInput(value));
+    assert.equal(value, color2);
 
     await vslsApi.end();
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsShareColor, null);
+    await updateLiveShareColor(LiveShareSettings.VSLSShareColor, '');
   });
 
   test('Workspace color is reverted after Live Share session when updated the color multiple times.', async () => {
@@ -215,35 +171,26 @@ suite('Live Share Integration', () => {
 
     const color2 = '#68217a';
     const fakeResponse2 = `C# Purple -> ${color2}`;
-    const stub2 = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse2));
+    const stub2 = await stubQuickPick(fakeResponse2);
 
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
-      LiveShareCommands.changeColorOfLiveShareHost,
-    );
+    await executeCommand(LiveShareCommands.changeColorOfLiveShareHost);
     stub2.restore();
 
     await timeout(1000);
 
     const color3 = '#b52e31';
-    const fakeResponse3 = `Abgular Red -> ${color3}`;
-    const stub3 = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse3));
+    const fakeResponse3 = `Angular Red -> ${color3}`;
+    const stub3 = await stubQuickPick(fakeResponse3);
 
-    await executeCommand<vscode.ExtensionContext>(LiveShareCommands.changeColorOfLiveShareHost);
+    await executeCommand(LiveShareCommands.changeColorOfLiveShareHost);
     stub3.restore();
 
     await timeout(1000);
 
     const color4 = '#639';
     const fakeResponse4 = `Gatsby Purple -> ${color4}`;
-    const stub4 = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse4));
-
-    await executeCommand<vscode.ExtensionContext>(LiveShareCommands.changeColorOfLiveShareHost);
+    const stub4 = await stubQuickPick(fakeResponse4);
+    await executeCommand(LiveShareCommands.changeColorOfLiveShareHost);
     stub4.restore();
 
     await timeout(1000);
@@ -255,10 +202,10 @@ suite('Live Share Integration', () => {
     let config = getPeacockWorkspaceConfig();
     const value = config[ColorSettings.titleBar_activeBackground] as string;
 
-    assert(!isValidColorInput(value));
-    assert(value == null);
+    assert.ok(!isValidColorInput(value));
+    assert.ok(!value);
 
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsShareColor, null);
+    await updateLiveShareColor(LiveShareSettings.VSLSShareColor, '');
   });
 
   test('Workspace color is reverted to a preset color after Live Share session when updated the color multiple times.', async () => {
@@ -270,10 +217,7 @@ suite('Live Share Integration', () => {
 
     const startColor = '#007fff';
     const fakeResponse = `Azure Blue -> ${startColor}`;
-    const stub = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse));
-
+    const stub = await stubQuickPick(fakeResponse);
     await executeCommand(Commands.changeColorToFavorite);
     stub.restore();
 
@@ -283,35 +227,24 @@ suite('Live Share Integration', () => {
 
     const color2 = '#68217a';
     const fakeResponse2 = `C# Purple -> ${color2}`;
-    const stub2 = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse2));
-
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
-      LiveShareCommands.changeColorOfLiveShareHost,
-    );
+    const stub2 = await stubQuickPick(fakeResponse2);
+    await executeCommand(LiveShareCommands.changeColorOfLiveShareHost);
     stub2.restore();
 
     await timeout(1000);
 
     const color3 = '#b52e31';
-    const fakeResponse3 = `Abgular Red -> ${color3}`;
-    const stub3 = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse3));
-
-    await executeCommand<vscode.ExtensionContext>(LiveShareCommands.changeColorOfLiveShareHost);
+    const fakeResponse3 = `Angular Red -> ${color3}`;
+    const stub3 = await stubQuickPick(fakeResponse3);
+    await executeCommand(LiveShareCommands.changeColorOfLiveShareHost);
     stub3.restore();
 
     await timeout(1000);
 
     const color4 = '#639';
     const fakeResponse4 = `Gatsby Purple -> ${color4}`;
-    const stub4 = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse4));
-
-    await executeCommand<vscode.ExtensionContext>(LiveShareCommands.changeColorOfLiveShareHost);
+    const stub4 = await stubQuickPick(fakeResponse4);
+    await executeCommand(LiveShareCommands.changeColorOfLiveShareHost);
     stub4.restore();
 
     await timeout(1000);
@@ -323,9 +256,12 @@ suite('Live Share Integration', () => {
     let config = getPeacockWorkspaceConfig();
     const value = config[ColorSettings.titleBar_activeBackground] as string;
 
-    assert(isValidColorInput(value));
-    assert(value === startColor);
+    assert.ok(isValidColorInput(value));
+    assert.equal(value, startColor);
 
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsShareColor, null);
+    await updateLiveShareColor(LiveShareSettings.VSLSShareColor, '');
   });
 });
+
+const stubQuickPick = async (fakeResponse: string) =>
+  await sinon.stub(vscode.window, 'showQuickPick').returns(Promise.resolve<any>(fakeResponse));