+++ src/mementos.ts
@@ -1,8 +1,6 @@
 import { isValidColorInput } from './color-library';
 import { peacockMementos, extensionShortName, State } from './models';
 import { Logger } from './logging';
-import { peacockVslsMementos } from './live-share/constants';
-import { peacockRemoteMementos } from './remote/constants';
 
 export interface IMementoLog {
   name: string;
@@ -55,11 +53,6 @@ export async function resetMementos() {
 
   // Global
   await ec.globalState.update(peacockMementos.favoritesVersion, undefined);
-  await ec.globalState.update(peacockVslsMementos.vslsJoinColor, undefined);
-  await ec.globalState.update(peacockVslsMementos.vslsShareColor, undefined);
-  await ec.globalState.update(peacockRemoteMementos.remoteContainersColor, undefined);
-  await ec.globalState.update(peacockRemoteMementos.remoteSshColor, undefined);
-  await ec.globalState.update(peacockRemoteMementos.remoteWslColor, undefined);
 
   // Workspace
   await ec.workspaceState.update(peacockMementos.peacockColor, undefined);
@@ -74,31 +67,6 @@ export function getMementos() {
     name: peacockMementos.favoritesVersion,
     type: 'globalState',
     value: ec.globalState.get(peacockMementos.favoritesVersion),
-  });
-  mementos.push({
-    name: peacockVslsMementos.vslsJoinColor,
-    type: 'globalState',
-    value: ec.globalState.get(peacockVslsMementos.vslsJoinColor),
-  });
-  mementos.push({
-    name: peacockVslsMementos.vslsShareColor,
-    type: 'globalState',
-    value: ec.globalState.get(peacockVslsMementos.vslsShareColor),
-  });
-  mementos.push({
-    name: peacockRemoteMementos.remoteContainersColor,
-    type: 'globalState',
-    value: ec.globalState.get(peacockRemoteMementos.remoteContainersColor),
-  });
-  mementos.push({
-    name: peacockRemoteMementos.remoteSshColor,
-    type: 'globalState',
-    value: ec.globalState.get(peacockRemoteMementos.remoteSshColor),
-  });
-  mementos.push({
-    name: peacockRemoteMementos.remoteWslColor,
-    type: 'globalState',
-    value: ec.globalState.get(peacockRemoteMementos.remoteWslColor),
   });
 
   // Workspace